pyproject.toml
[project]
name = "billmunshi"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0,<6.0",
    "django-cors-headers>=4.3.0",
    "django-environ>=0.11.0",
    "django-filter>=24.2",
    "django-redis>=5.4.0",
    "django-waffle>=5.0.0",
    "djangorestframework>=3.15.0",
    "djangorestframework-api-key>=3.0.0",
    "djangorestframework-simplejwt>=5.3.0",
    "drf-spectacular>=0.27.0",
    "pdf2image>=1.17.0",
    "psycopg2-binary>=2.9.10",
    "psycopg[binary]>=3.1.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
